+++ server/internal/common/dto/pagination.go
@@ -1,8 +1,8 @@
 package dto
 
 type PaginationQuery struct {
-	Page       *int32  `query:"page" validate:"omitempty,gte=0,lte=100"`
-	Limit      *int32  `query:"limit" validate:"omitempty,gte=1,lte=100"`
+	Page       *int64  `query:"page" validate:"omitempty,gte=0,lte=100"`
+	Limit      *int64  `query:"limit" validate:"omitempty,gte=1,lte=100"`
 	NextCursor *string `query:"cursor"`
 }
 
@@ -11,12 +11,12 @@ func (p *PaginationQuery) UseCursor() bool {
 }
 
 func (p *PaginationQuery) SetDefaults() {
-	defaultLimit := int32(10)
+	defaultLimit := int64(10)
 	if p.Limit == nil {
 		p.Limit = &defaultLimit
 	}
 	if p.Page == nil {
-		defaultPage := int32(1)
+		defaultPage := int64(1)
 		p.Page = &defaultPage
 	}
 }
