+++ server/internal/common/helpers/pagination.go
@@ -2,7 +2,7 @@ package helpers
 
 import "ChainServer/internal/common/response"
 
-func BuildPaginationMeta(limit, page, total int32, nextCursor any) *response.PaginationMeta {
+func BuildPaginationMeta(limit, page, total int64, nextCursor any) *response.PaginationMeta {
 	return &response.PaginationMeta{
 		Limit:       int(limit),
 		CurrentPage: int(page),
